langeval-core/resource-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.8
psycopg2-binary>=2.9.0
pydantic-settings>=2.0.0
httpx>=0.25.0
cryptography>=41.0.0
clickhouse-connect>=0.6.0
redis>=5.0.0
openai>=1.0.0